requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
MySQL-python==1.2.4
SQLAlchemy==0.8.2
Werkzeug==0.9.4
argparse==1.2.1
itsdangerous==0.23
python-dateutil==2.1
requests==2.0.0
six==1.4.1
wsgiref==0.1.2
